+++ src/azureExtensions.ts
@@ -4,8 +4,31 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { AppResource } from "@microsoft/vscode-azext-utils/hostapi";
+import { extensions } from "vscode";
 import { AzExtResourceType } from "../api/src/index";
 import { localize } from "./utils/localize";
+
+
+/**
+ * This is a temporary function usend to optionaly enable support for MongoClusters in the Azure Resources extension.
+ *
+ * This solution is necessary for a staged release of the MongoClusters feature from the vscode-cosmosdb extension.
+ * It will be removed once the MongoClusters feature is fully released.
+ *
+ * @returns
+ */
+function enableMongoClustersSupport() {
+    const vsCodeCosmosDBConfiguration = extensions.getExtension('ms-azuretools.vscode-cosmosdb')?.packageJSON as ExtensionPackageMongoClustersEnabled;
+    return (vsCodeCosmosDBConfiguration && vsCodeCosmosDBConfiguration.enableMongoClusters);
+}
+
+/**
+ * This is a temporary interface used to enable support for MongoClusters in the Azure Resources extension.
+ * It will be removed once the MongoClusters feature is fully released.
+ */
+interface ExtensionPackageMongoClustersEnabled {
+    readonly enableMongoClusters?: boolean;
+}
 
 export const azureExtensions: IAzExtMetadata[] = [
     {
@@ -77,11 +100,23 @@ export const azureExtensions: IAzExtMetadata[] = [
     {
         name: 'vscode-cosmosdb',
         label: 'Databases',
-        resourceTypes: [
-            AzExtResourceType.AzureCosmosDb,
-            AzExtResourceType.PostgresqlServersStandard,
-            AzExtResourceType.PostgresqlServersFlexible,
-        ],
+        resourceTypes:
+            /**
+            * This is a temporary interface used to enable support for MongoClusters in the Azure Resources extension.
+            * It will be removed once the MongoClusters feature is fully released.
+            */
+            enableMongoClustersSupport() ?
+                [
+                    AzExtResourceType.AzureCosmosDb,
+                    AzExtResourceType.MongoClusters,
+                    AzExtResourceType.PostgresqlServersStandard,
+                    AzExtResourceType.PostgresqlServersFlexible,
+                ] :
+                [
+                    AzExtResourceType.AzureCosmosDb,
+                    AzExtResourceType.PostgresqlServersStandard,
+                    AzExtResourceType.PostgresqlServersFlexible,
+                ],
         reportIssueCommandId: 'azureDatabases.reportIssue'
     },
     {
@@ -137,6 +172,7 @@ export const legacyTypeMap: Partial<Record<AzExtResourceType, string>> = {
     StaticWebApps: 'microsoft.web/staticsites',
     VirtualMachines: 'microsoft.compute/virtualmachines',
     AzureCosmosDb: 'microsoft.documentdb/databaseaccounts',
+    MongoClusters: 'microsoft.documentdb/mongoclusters',
     PostgresqlServersStandard: 'microsoft.dbforpostgresql/servers',
     PostgresqlServersFlexible: 'microsoft.dbforpostgresql/flexibleservers',
     SpringApps: 'microsoft.appplatform/spring',