+++ src/commands/revealResource.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ParsedAzureResourceId } from '@microsoft/vscode-azext-azureutils';
-import { AzExtTreeItem, IActionContext, parseError } from '@microsoft/vscode-azext-utils';
+import { AzExtTreeItem, IActionContext, maskUserInfo, parseError } from '@microsoft/vscode-azext-utils';
 import { VSCodeRevealOptions } from '../../api/src/index';
 import { ext } from '../extensionVariables';
 import { ResourceGroupsItem } from '../tree/ResourceGroupsItem';
@@ -19,7 +19,7 @@ export async function revealResource(context: IActionContext, resourceId: string
             await ext.appResourceTreeView.reveal(item as unknown as AzExtTreeItem, options ?? { expand: false, focus: true, select: true });
         }
     } catch (error) {
-        context.telemetry.properties.revealError = parseError(error).message;
+        context.telemetry.properties.revealError = maskUserInfo(parseError(error).message, []);
     }
 }